+++ NET-Core/LibUA/NetDispatcher.cs
@@ -1508,7 +1508,7 @@ namespace LibUA
                 if (!recvBuf.Decode(out uint securityTokenRequestType)) { return ErrorParseFail; }
                 if (!recvBuf.Decode(out uint messageSecurityMode)) { return ErrorParseFail; }
                 if (!recvBuf.DecodeUAByteString(out byte[] clientNonce)) { return ErrorParseFail; }
-                if (!recvBuf.Decode(out uint reqLifetime)) { return ErrorParseFail; }
+                if (!recvBuf.Decode(out double reqLifetime)) { return ErrorParseFail; }
 
                 try
                 {
@@ -1572,7 +1572,7 @@ namespace LibUA
                     throw new Exception();
                 }
 
-                config.TokenLifetime = Math.Min(reqLifetime, MaxTokenLifetime);
+                config.TokenLifetime = Math.Min((uint)reqLifetime, MaxTokenLifetime);
                 config.TokenCreatedAt = reqHeader.Timestamp;
 
                 if (config.SecurityPolicy == SecurityPolicy.None)
@@ -1674,7 +1674,7 @@ namespace LibUA
                 succeeded &= respBuf.Encode(config.ChannelID);
                 succeeded &= respBuf.Encode(config.TokenID);
                 succeeded &= respBuf.Encode((UInt64)config.TokenCreatedAt.ToFileTime());
-                succeeded &= respBuf.Encode(config.TokenLifetime);
+                succeeded &= respBuf.Encode((double)config.TokenLifetime);
 
                 succeeded &= respBuf.EncodeUAByteString(config.LocalNonce);
